refactor: staticize entry callout
