+++ gql/gql.ts
@@ -64,9 +64,13 @@ const documents = {
     types.AccordionGroupBlockFragmentDoc,
   'fragment ctaGridBlock on contentBlocks_ctaGrid_BlockType {\n  typeHandle\n  id\n  header\n  backgroundColor\n  fullWidth\n  mixedLink {\n    ...MixedLinkFragment\n  }\n  items: children {\n    ... on contentBlocks_cta_BlockType {\n      id\n      contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 572, height: 316)\n        }\n      }\n      mixedLink {\n        ...MixedLinkFragment\n      }\n    }\n  }\n}':
     types.CtaGridBlockFragmentDoc,
+  'fragment CalloutEvent on callouts_calloutEvent_Entry {\n  calloutType: typeHandle\n  backgroundColor\n  id\n  entry: eventEntry {\n    ... on events_events_Entry {\n      title\n      url\n      startDate\n      endDate: date\n      description\n      hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      image: hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      entryType: eventType {\n        title\n        slug\n      }\n    }\n  }\n}':
+    types.CalloutEventFragmentDoc,
   "fragment CalloutMain on callouts_callout_Entry {\n  backgroundColor\n  dynamicComponent\n  header\n  cantoAssetSingle {\n    ...CantoAssetMetadata\n  }\n  links {\n    ...LinksFragment\n  }\n  padImage\n  order\n  ratio\n  text\n  calloutType: typeHandle\n  width\n}":
     types.CalloutMainFragmentDoc,
-  'fragment calloutBlock on contentBlocks_callout_BlockType {\n  id\n  typeHandle\n  callout {\n    ...CalloutMain\n    ... on callouts_calloutTwoTone_Entry {\n      backgroundColor\n      backgroundImage: cantoAssetSingle {\n        ...CantoAssetSingle\n      }\n      captionRichText\n      header\n      image: contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 800, height: 550)\n        }\n      }\n      links {\n        ...LinksFragment\n      }\n      text\n      calloutType: typeHandle\n    }\n    ... on callouts_calloutQuote_Entry {\n      calloutType: typeHandle\n      header\n      text\n      colorScheme\n      links {\n        ...LinksFragment\n      }\n      quote\n      attribution\n      imageQuote {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 492, height: 492)\n        }\n      }\n    }\n    ... on callouts_calloutNews_Entry {\n      calloutType: typeHandle\n      backgroundColor\n      id\n      entry: newsEntry {\n        ... on news_post_Entry {\n          title\n          url\n          date\n          dateCreated\n          description: teaser\n          pressReleaseId\n          hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          image: hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          entryType: postType {\n            title\n            slug\n          }\n          postType {\n            id\n            title\n            slug\n          }\n        }\n      }\n    }\n    ... on callouts_calloutEvent_Entry {\n      calloutType: typeHandle\n      backgroundColor\n      id\n      entry: eventEntry {\n        ... on events_events_Entry {\n          title\n          url\n          startDate\n          endDate: date\n          description\n          hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          image: hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          entryType: eventType {\n            title\n            slug\n          }\n        }\n      }\n    }\n  }\n}':
+  'fragment CalloutNews on callouts_calloutNews_Entry {\n  calloutType: typeHandle\n  backgroundColor\n  id\n  entry: newsEntry {\n    ... on news_post_Entry {\n      title\n      url\n      date\n      dateCreated\n      description: teaser\n      pressReleaseId\n      hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      image: hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      entryType: postType {\n        title\n        slug\n      }\n      postType {\n        id\n        title\n        slug\n      }\n    }\n  }\n}':
+    types.CalloutNewsFragmentDoc,
+  'fragment calloutBlock on contentBlocks_callout_BlockType {\n  id\n  typeHandle\n  callout {\n    ...CalloutMain\n    ... on callouts_calloutTwoTone_Entry {\n      backgroundColor\n      backgroundImage: cantoAssetSingle {\n        ...CantoAssetSingle\n      }\n      captionRichText\n      header\n      image: contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 800, height: 550)\n        }\n      }\n      links {\n        ...LinksFragment\n      }\n      text\n      calloutType: typeHandle\n    }\n    ... on callouts_calloutQuote_Entry {\n      calloutType: typeHandle\n      header\n      text\n      colorScheme\n      links {\n        ...LinksFragment\n      }\n      quote\n      attribution\n      imageQuote {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 492, height: 492)\n        }\n      }\n    }\n    ...CalloutNews\n    ...CalloutEvent\n  }\n}':
     types.CalloutBlockFragmentDoc,
   "fragment complexTableBlock on contentBlocks_complexTable_BlockType {\n  id\n  typeHandle\n  plainText\n  verticalAlignment\n  sites\n  complexTable {\n    ... on complexTable_BlockType {\n      tableRow {\n        ... on tableRow_tableCell_BlockType {\n          id\n          cellWidth\n          cellBackground\n          cellContent\n          hasFlexibleCellWidth\n        }\n      }\n    }\n  }\n}":
     types.ComplexTableBlockFragmentDoc,
@@ -352,14 +356,26 @@ export function graphql(
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
+  source: 'fragment CalloutEvent on callouts_calloutEvent_Entry {\n  calloutType: typeHandle\n  backgroundColor\n  id\n  entry: eventEntry {\n    ... on events_events_Entry {\n      title\n      url\n      startDate\n      endDate: date\n      description\n      hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      image: hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      entryType: eventType {\n        title\n        slug\n      }\n    }\n  }\n}'
+): (typeof documents)['fragment CalloutEvent on callouts_calloutEvent_Entry {\n  calloutType: typeHandle\n  backgroundColor\n  id\n  entry: eventEntry {\n    ... on events_events_Entry {\n      title\n      url\n      startDate\n      endDate: date\n      description\n      hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      image: hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      entryType: eventType {\n        title\n        slug\n      }\n    }\n  }\n}'];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(
   source: "fragment CalloutMain on callouts_callout_Entry {\n  backgroundColor\n  dynamicComponent\n  header\n  cantoAssetSingle {\n    ...CantoAssetMetadata\n  }\n  links {\n    ...LinksFragment\n  }\n  padImage\n  order\n  ratio\n  text\n  calloutType: typeHandle\n  width\n}"
 ): (typeof documents)["fragment CalloutMain on callouts_callout_Entry {\n  backgroundColor\n  dynamicComponent\n  header\n  cantoAssetSingle {\n    ...CantoAssetMetadata\n  }\n  links {\n    ...LinksFragment\n  }\n  padImage\n  order\n  ratio\n  text\n  calloutType: typeHandle\n  width\n}"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
-  source: 'fragment calloutBlock on contentBlocks_callout_BlockType {\n  id\n  typeHandle\n  callout {\n    ...CalloutMain\n    ... on callouts_calloutTwoTone_Entry {\n      backgroundColor\n      backgroundImage: cantoAssetSingle {\n        ...CantoAssetSingle\n      }\n      captionRichText\n      header\n      image: contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 800, height: 550)\n        }\n      }\n      links {\n        ...LinksFragment\n      }\n      text\n      calloutType: typeHandle\n    }\n    ... on callouts_calloutQuote_Entry {\n      calloutType: typeHandle\n      header\n      text\n      colorScheme\n      links {\n        ...LinksFragment\n      }\n      quote\n      attribution\n      imageQuote {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 492, height: 492)\n        }\n      }\n    }\n    ... on callouts_calloutNews_Entry {\n      calloutType: typeHandle\n      backgroundColor\n      id\n      entry: newsEntry {\n        ... on news_post_Entry {\n          title\n          url\n          date\n          dateCreated\n          description: teaser\n          pressReleaseId\n          hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          image: hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          entryType: postType {\n            title\n            slug\n          }\n          postType {\n            id\n            title\n            slug\n          }\n        }\n      }\n    }\n    ... on callouts_calloutEvent_Entry {\n      calloutType: typeHandle\n      backgroundColor\n      id\n      entry: eventEntry {\n        ... on events_events_Entry {\n          title\n          url\n          startDate\n          endDate: date\n          description\n          hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          image: hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          entryType: eventType {\n            title\n            slug\n          }\n        }\n      }\n    }\n  }\n}'
-): (typeof documents)['fragment calloutBlock on contentBlocks_callout_BlockType {\n  id\n  typeHandle\n  callout {\n    ...CalloutMain\n    ... on callouts_calloutTwoTone_Entry {\n      backgroundColor\n      backgroundImage: cantoAssetSingle {\n        ...CantoAssetSingle\n      }\n      captionRichText\n      header\n      image: contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 800, height: 550)\n        }\n      }\n      links {\n        ...LinksFragment\n      }\n      text\n      calloutType: typeHandle\n    }\n    ... on callouts_calloutQuote_Entry {\n      calloutType: typeHandle\n      header\n      text\n      colorScheme\n      links {\n        ...LinksFragment\n      }\n      quote\n      attribution\n      imageQuote {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 492, height: 492)\n        }\n      }\n    }\n    ... on callouts_calloutNews_Entry {\n      calloutType: typeHandle\n      backgroundColor\n      id\n      entry: newsEntry {\n        ... on news_post_Entry {\n          title\n          url\n          date\n          dateCreated\n          description: teaser\n          pressReleaseId\n          hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          image: hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          entryType: postType {\n            title\n            slug\n          }\n          postType {\n            id\n            title\n            slug\n          }\n        }\n      }\n    }\n    ... on callouts_calloutEvent_Entry {\n      calloutType: typeHandle\n      backgroundColor\n      id\n      entry: eventEntry {\n        ... on events_events_Entry {\n          title\n          url\n          startDate\n          endDate: date\n          description\n          hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          image: hero {\n            ... on heroes_Asset {\n              altText\n              width\n              height\n              url @transform(mode: "crop", width: 800, height: 550)\n            }\n          }\n          entryType: eventType {\n            title\n            slug\n          }\n        }\n      }\n    }\n  }\n}'];
+  source: 'fragment CalloutNews on callouts_calloutNews_Entry {\n  calloutType: typeHandle\n  backgroundColor\n  id\n  entry: newsEntry {\n    ... on news_post_Entry {\n      title\n      url\n      date\n      dateCreated\n      description: teaser\n      pressReleaseId\n      hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      image: hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      entryType: postType {\n        title\n        slug\n      }\n      postType {\n        id\n        title\n        slug\n      }\n    }\n  }\n}'
+): (typeof documents)['fragment CalloutNews on callouts_calloutNews_Entry {\n  calloutType: typeHandle\n  backgroundColor\n  id\n  entry: newsEntry {\n    ... on news_post_Entry {\n      title\n      url\n      date\n      dateCreated\n      description: teaser\n      pressReleaseId\n      hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      image: hero {\n        ... on heroes_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 900, height: 550)\n        }\n      }\n      entryType: postType {\n        title\n        slug\n      }\n      postType {\n        id\n        title\n        slug\n      }\n    }\n  }\n}'];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(
+  source: 'fragment calloutBlock on contentBlocks_callout_BlockType {\n  id\n  typeHandle\n  callout {\n    ...CalloutMain\n    ... on callouts_calloutTwoTone_Entry {\n      backgroundColor\n      backgroundImage: cantoAssetSingle {\n        ...CantoAssetSingle\n      }\n      captionRichText\n      header\n      image: contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 800, height: 550)\n        }\n      }\n      links {\n        ...LinksFragment\n      }\n      text\n      calloutType: typeHandle\n    }\n    ... on callouts_calloutQuote_Entry {\n      calloutType: typeHandle\n      header\n      text\n      colorScheme\n      links {\n        ...LinksFragment\n      }\n      quote\n      attribution\n      imageQuote {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 492, height: 492)\n        }\n      }\n    }\n    ...CalloutNews\n    ...CalloutEvent\n  }\n}'
+): (typeof documents)['fragment calloutBlock on contentBlocks_callout_BlockType {\n  id\n  typeHandle\n  callout {\n    ...CalloutMain\n    ... on callouts_calloutTwoTone_Entry {\n      backgroundColor\n      backgroundImage: cantoAssetSingle {\n        ...CantoAssetSingle\n      }\n      captionRichText\n      header\n      image: contentImage {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 800, height: 550)\n        }\n      }\n      links {\n        ...LinksFragment\n      }\n      text\n      calloutType: typeHandle\n    }\n    ... on callouts_calloutQuote_Entry {\n      calloutType: typeHandle\n      header\n      text\n      colorScheme\n      links {\n        ...LinksFragment\n      }\n      quote\n      attribution\n      imageQuote {\n        ... on contentImages_Asset {\n          altText\n          width\n          height\n          url @transform(mode: "crop", width: 492, height: 492)\n        }\n      }\n    }\n    ...CalloutNews\n    ...CalloutEvent\n  }\n}'];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
