+++ helpers/styles.ts
@@ -1,3 +1,7 @@
 export const isDarkMode = (backgroundColor?: string | null) => {
-  return !!backgroundColor?.includes("invert");
+  if (!backgroundColor) return false;
+
+  return (
+    !!backgroundColor.includes("invert") || !!backgroundColor.includes("black")
+  );
 };
