+++ lib/api/noirlabReleases.js
@@ -1,8 +1,7 @@
 import useSWR from "swr";
 import { useMemo } from "react";
 import { getLocaleString } from "@/lib/helpers/site";
-import { makeReleaseFeature } from "@/lib/utils";
-import { NOIRLabServices } from "./noirlab";
+import { NOIRLabServices, makeReleaseFeature } from "./noirlab";
 
 export function useRelease(site, entry) {
   const locale = getLocaleString(site);
