+++ lib/utils.js
@@ -197,29 +197,6 @@ export function normalizeItemData(items, whichEntry = "entry") {
     .map((item) => item[whichEntry]?.[0]);
 }
 
-/** @function makeReleaseFeature
- * @param {Array<any>} images
- * @param {"banner1920" | "large" | "medium" | "news" | "newsfeature" | "newsmini" | "original" | "screen" | "screen640" | "thumb350x" | "thumb700x"} format
- */
-export function makeReleaseFeature(images, format = "screen") {
-  if (!images) return;
-  const feature = images[0];
-  if (!feature) return;
-
-  const { title, height, width, formats } = feature;
-
-  return [
-    {
-      altText: title,
-      url: formats[format],
-      url2x: formats[format],
-      url3x: formats[format],
-      height,
-      width,
-    },
-  ];
-}
-
 // CATEGORY STUFF
 export const getCategoryId = (categories, slug) => {
   return categories.reduce((a, b) => (b.slug === slug ? a + b.id : a), "");
