+++ services/craft/entries/content-blocks/CalloutEvent.gql
@@ -1,0 +1,34 @@
+fragment CalloutEvent on callouts_calloutEvent_Entry {
+  calloutType: typeHandle
+  backgroundColor
+  id
+  entry: eventEntry {
+    ... on events_events_Entry {
+      title
+      url
+      startDate
+      endDate: date
+      description
+      hero {
+        ... on heroes_Asset {
+          altText
+          width
+          height
+          url @transform(mode: "crop", width: 900, height: 550)
+        }
+      }
+      image: hero {
+        ... on heroes_Asset {
+          altText
+          width
+          height
+          url @transform(mode: "crop", width: 900, height: 550)
+        }
+      }
+      entryType: eventType {
+        title
+        slug
+      }
+    }
+  }
+}
