+++ services/craft/entries/content-blocks/Callouts.gql
@@ -43,79 +43,7 @@ fragment calloutBlock on contentBlocks_callout_BlockType {
         }
       }
     }
-    ... on callouts_calloutNews_Entry {
-      calloutType: typeHandle
-      backgroundColor
-      id
-      entry: newsEntry {
-        ... on news_post_Entry {
-          title
-          url
-          date
-          dateCreated
-          description: teaser
-          pressReleaseId
-          hero {
-            ... on heroes_Asset {
-              altText
-              width
-              height
-              url @transform(mode: "crop", width: 800, height: 550)
-            }
-          }
-          image: hero {
-            ... on heroes_Asset {
-              altText
-              width
-              height
-              url @transform(mode: "crop", width: 800, height: 550)
-            }
-          }
-          entryType: postType {
-            title
-            slug
-          }
-          postType {
-            id
-            title
-            slug
-          }
-        }
-      }
-    }
-    ... on callouts_calloutEvent_Entry {
-      calloutType: typeHandle
-      backgroundColor
-      id
-      entry: eventEntry {
-        ... on events_events_Entry {
-          title
-          url
-          startDate
-          endDate: date
-          description
-          hero {
-            ... on heroes_Asset {
-              altText
-              width
-              height
-              url @transform(mode: "crop", width: 800, height: 550)
-            }
-          }
-          image: hero {
-            ... on heroes_Asset {
-              altText
-              width
-              height
-              url @transform(mode: "crop", width: 800, height: 550)
-            }
-          }
-          entryType: eventType {
-            title
-            slug
-          }
-        }
-      }
-    }
+    ...CalloutNews
+    ...CalloutEvent
   }
 }
